Factor out autocomplete.js.
